src/main/java/de/neuefische/studentdbweb/repository/StudentRepository.java
package de.neuefische.studentdbweb.repository;

import de.neuefische.studentdbweb.model.Student;
import org.springframework.stereotype.Repository;

import java.util.ArrayList;
import java.util.List;
import java.util.Optional;

@Repository
public class StudentRepository {

    private final List<Student> students = new ArrayList<>();

    public List<Student> list() {
        return students;
    }

    public Optional<Student> findById(String id) {
        for (Student student : students) {
            if(student.getId().equals(id)){
                return Optional.of(student);
            }
        }
        return Optional.empty();
    }


    public Student addStudent(Student student){
        this.students.add(student);
        return student;
    }
}
